main

+++ struct.go
@@ -96,9 +96,9 @@ type GeoJsonPolygon struct {
 
 type Point struct {
 	Coordinates []float64 `json:"coordinates" bson:"coordinates"`
-	Max         int64     `json:"max,omitempty" bson:"max,omitempty"`
-	Min         int64     `json:"min,omitempty" bson:"min,omitempty"`
-	Radius      float64     `json:"radius,omitempty" bson:"radius,omitempty"`
+	Max         float64   `json:"max,omitempty" bson:"max,omitempty"`
+	Min         float64   `json:"min,omitempty" bson:"min,omitempty"`
+	Radius      float64   `json:"radius,omitempty" bson:"radius,omitempty"`
 }
 
 type Polyline struct {
